+++ changelog.md
@@ -2,9 +2,12 @@ Codion Change Log
 ==================
 
 ## 0.18.2
+### is.codion.swing.common.model
+- DefaultFilterTableModel now validates column identifier uniqueness.
 ### is.codion.swing.common.ui
 - Control.control() renamed commandControl().
 - Control.BuilderFactory.actionCommand() renamed action().
+- FilterTable now validates column identifier uniqueness.
 ### is.codion.framework.domain
 - DefaultColumnDefinition.groupBy, aggregate, selectable and lazy no longer transient.
 ### is.codion.framework.model